+++ aiohttp/worker.py
@@ -241,7 +241,7 @@ class WSGIServer(Server):
             loop=self._loop,
             logger=self.worker.log,
             debug=self.worker.log.loglevel == logging.DEBUG,
-            keep_alive=self.worker.cfg.keepalive,
+            keepalive_timeout=self.worker.cfg.keepalive,
             access_log=self.worker.log.access_log,
             access_log_format=self.access_log_format)